+++ src/app.ts
@@ -21,6 +21,7 @@ import manageRoutes from "./routes/supplierRoutes";
 import cookieParser from "cookie-parser";
 import bodyParser from "body-parser";
 import session from "express-session";
+import { backLink } from "./middleware/backMiddleware";
 import { handleCookies } from "./middleware/cookieMiddleware";
 import passport from "passport";
 import {
@@ -112,30 +113,8 @@ env.addFilter("formatDate", function (date: string | number | Date) {
 
 // Set Nunjucks as the Express view engine
 app.set("view engine", "njk");
-app.use((req, res, next) => {
-  const fromBack = req.query.fromBack === "true";
-  req.session.pageHistory = req.session.pageHistory || [];
-  const urlIndex = req.session.pageHistory.indexOf(req.url);
 
-  if (
-    req.method === "GET" &&
-    !fromBack &&
-    !req.session.pageHistory.includes(req.url)
-  ) {
-    req.session.pageHistory.push(req.url);
-    const maxHistoryLength = 5;
-    if (req.session.pageHistory.length > maxHistoryLength) {
-      req.session.pageHistory.shift();
-    }
-    // res.locals.pageHistory = req.session.pageHistory;
-    // console.log(res.locals.pageHistory);
-  } else {
-    req.session.pageHistory.splice(urlIndex, 1);
-  }
-  res.locals.pageHistory = req.session.pageHistory;
-  console.log(res.locals.pageHistory);
-  next();
-});
+app.use(backLink);
 
 app.use("/", loginRoutes);
 app.use("/auth", authRoutes);
